Initial commit, moved NgsaLogger to Core library
