+++ src/LodeRunner.Core/NgsaLogger/NgsaLoggerExtensions.cs
@@ -4,7 +4,7 @@
 using System;
 using Microsoft.Extensions.Logging;
 
-namespace LodeRunner.API.Middleware
+namespace LodeRunner.Core.NgsaLogger
 {
     /// <summary>
     /// NGSA Logger Extensions.
