+++ src/LodeRunner.Core/NgsaLogger/NgsaLoggerProvider.cs
@@ -4,7 +4,7 @@
 using System.Collections.Concurrent;
 using Microsoft.Extensions.Logging;
 
-namespace LodeRunner.API.Middleware
+namespace LodeRunner.Core.NgsaLogger
 {
     /// <summary>
     /// NGSA Logger Provider.
